refactoring verify token file
