+++ server/middlewares/auth/index.js
@@ -1,19 +1,15 @@
-const { verify } = require('jsonwebtoken');
-
-const { SECRET_KEY } = process.env;
+const Boom = require('@hapi/boom');
+const { verifyToken } = require('../../utils');
 
 const Auth = async (req, res, next) => {
-  if (req.cookies.token) {
-    verify(req.cookies.token, SECRET_KEY, (err, payload) => {
-      if (err) {
-        res
-          .status(401)
-          .json({ statusCode: 401, message: 'You are not Authorized' });
-      } else {
-        req.user = payload;
-        next();
-      }
-    });
+  const { token } = req.cookies;
+  if (token) {
+    try {
+      req.user = await verifyToken(token);
+      next();
+    } catch (err) {
+      throw Boom.unauthorized('You are not Authorized');
+    }
   } else {
     res
       .status(401)
