+++ server/utils/index.js
@@ -1,8 +1,9 @@
 const { loginSchema, signupSchema } = require('./validation');
-const signToken = require('./signToken');
+const { signToken, verifyToken } = require('./jwt');
 
 module.exports = {
   loginSchema,
   signupSchema,
   signToken,
+  verifyToken,
 };
